+++ lib/src/common/ubi_mem.c
@@ -11,6 +11,7 @@
 /* Internal headers: */
 #include "ubi_mem.h"
 #include "ubi_internal.h"
+#include "ubi_io.h"
 #include "ubi_cache.h"
 
 /* Public headers: */